objects: move SkidooArmed_Setup to its own file
