+++ data/tr2/ship/cfg/TR2X_gameflow.json5
@@ -135,7 +135,7 @@
         "O_BANDIT_2": "Mercenary 2",
         "O_BANDIT_2B": "Mercenary 3",
         "O_SKIDOO_ARMED": "Black Snowmobile",
-        "O_SKIDMAN": "Black Snowmobile Driver",
+        "O_SKIDOO_DRIVER": "Black Snowmobile Driver",
         "O_MONK_1": "Monk 1",
         "O_MONK_2": "Monk 2",
         "O_FALLING_BLOCK_1": "Falling Block 1",
