+++ src/libtrx/include/libtrx/game/objects/ids_tr2.def
@@ -50,7 +50,7 @@ OBJ_ID_DEFINE(O_BANDIT_1, 48)
 OBJ_ID_DEFINE(O_BANDIT_2, 49)
 OBJ_ID_DEFINE(O_BANDIT_2B, 50)
 OBJ_ID_DEFINE(O_SKIDOO_ARMED, 51)
-OBJ_ID_DEFINE(O_SKIDMAN, 52)
+OBJ_ID_DEFINE(O_SKIDOO_DRIVER, 52)
 OBJ_ID_DEFINE(O_MONK_1, 53)
 OBJ_ID_DEFINE(O_MONK_2, 54)
 OBJ_ID_DEFINE(O_FALLING_BLOCK_1, 55)
