+++ src/libtrx/include/libtrx/game/objects/names_tr2.def
@@ -50,7 +50,7 @@ OBJ_NAME_DEFINE(O_BANDIT_1, "Mercenary 1")
 OBJ_NAME_DEFINE(O_BANDIT_2, "Mercenary 2")
 OBJ_NAME_DEFINE(O_BANDIT_2B, "Mercenary 3")
 OBJ_NAME_DEFINE(O_SKIDOO_ARMED, "Black Snowmobile")
-OBJ_NAME_DEFINE(O_SKIDMAN, "Black Snowmobile Driver")
+OBJ_NAME_DEFINE(O_SKIDOO_DRIVER, "Black Snowmobile Driver")
 OBJ_NAME_DEFINE(O_MONK_1, "Monk 1")
 OBJ_NAME_DEFINE(O_MONK_2, "Monk 2")
 OBJ_NAME_DEFINE(O_FALLING_BLOCK_1, "Falling Block 1")
