+++ src/tr2/game/objects/vehicles/skidoo_armed.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void SkidooArmed_Setup(void);
