+++ src/tr2/global/funcs.h
@@ -173,8 +173,8 @@
 #define GetCollisionAnim ((int32_t __cdecl (*)(ITEM *skidoo, XYZ_32 *moved))0x0043D600)
 #define SkidooDriver_Initialise ((void __cdecl (*)(int16_t item_num))0x0043ED40)
 #define SkidooDriver_Control ((void __cdecl (*)(int16_t rider_num))0x0043EDD0)
-#define SkidooDriver_Push ((void __cdecl (*)(ITEM *item, const ITEM *lara_item, int32_t radius))0x0043F1D0)
-#define SkidooDriver_Collision ((void __cdecl (*)(int16_t item_num, ITEM *lara_item, COLL_INFO *coll))0x0043F2F0)
+#define SkidooArmed_Push ((void __cdecl (*)(ITEM *item, const ITEM *lara_item, int32_t radius))0x0043F1D0)
+#define SkidooArmed_Collision ((void __cdecl (*)(int16_t item_num, ITEM *lara_item, COLL_INFO *coll))0x0043F2F0)
 #define Music_GetRealTrack ((int32_t __cdecl (*)(int32_t track))0x0043F380)
 #define Collide_TestCollision ((int32_t __cdecl (*)(ITEM *item, const ITEM *lara_item))0x0043F9B0)
 #define Collide_GetSpheres ((int32_t __cdecl (*)(const ITEM *item, SPHERE *spheres, bool world_space))0x0043FAE0)
